requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
pyjwt
requests